Benchmark lambda lifting (#1327)
* Move pen benchmarks

* Add lambda lifting benchmark

* Fix Rust benchmarks

* Fix benchmark script

* Refactor
