+++ benchmark/pen/lambda-lifting/main.pen
@@ -1,0 +1,22 @@
+import Core'Number
+import Os'File
+
+main = \(ctx context) none {
+  _ = File'Write(
+    ctx.Os,
+    File'StdOut(),
+    Number'String(sum(0, 100000000)),
+  )
+
+  none
+}
+
+sum = \(x number, i number) number {
+  add = \(x number, y number) number { x + y }
+
+  if i == 0 {
+    x
+  } else {
+    sum(add(x, i), i - 1)
+  }
+}
