+++ benchmark/pen/sum/pen.json
@@ -1,0 +1,7 @@
+{
+  "type": "application",
+  "dependencies": {
+    "Core": "pen:///core",
+    "Os": "pen:///os"
+  }
+}
