--- benchmark/rust/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "benchmark"
-version = "0.1.0"
-edition = "2021"
-
-[dev-dependencies]
-criterion = "0.3.6"
-im = "15.1.0"
-ordered-float = "3.0.0"
-
-[[bench]]
-name = "fibonacci"
-harness = false
-
-[[bench]]
-name = "hash_map"
-harness = false
-
-[[bench]]
-name = "sum"
-harness = false
