+++ benchmark/rust/fibonacci/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "fibonacci"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
