+++ benchmark/rust/im_hash_map_update/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "im-hash-map-update"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+im = "15"
+ordered-float = "3"
