--- benchmark/rust/src/main.rs
@@ -1,1 +1,0 @@
-fn main() {}
