+++ benchmark/rust/std_hash_map_insert/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "std-hash-map-insert"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+ordered-float = "3"
