+++ benchmark/rust/sum/src/main.rs
@@ -1,0 +1,9 @@
+fn main() {
+    let mut sum = 0.0;
+
+    for i in 0..=100_000_000 {
+        sum += i as f64;
+    }
+
+    dbg!(sum);
+}
